make changing start command failures clearer
The test failed with the start command not changing. This can happen if the
original stop message is missed.

This will make it clear if the stop message was never received, as the test
will fail to see '404', and continue seeing 'foo'.

Also, give the app more time to come up.