Revert "Manage spaces on a per-node basis"
This reverts commit c39c732d04c7f9a2f7e811a4192df0b99ddbf0a2.

The quantity of `gcf auth` calls that the space-managing CATs make in
a parallel run is unsustainable, and may cause gcf commands to take more
than 10 seconds. At this point the cmdtest matchers will report that
they have failed, failing the test suite. We're rolling back this
change to revisit our space-management approach.