Make staging logs test a bit less flakey
This was previously asserting that the LAST line of output from the staging process
would be seen in the CLI's output, but this is not guaranteed based on how v6.0
reads logs. As soon as the first app instance is up, we stop reading and printing
staging logs. With that in mind, it made more sense to write a matcher that would
attempt to match one of several staging log lines.