tests set up and tear down organization
This no longer has to be configured. Also, all of this can now be specified in
the config .json files (though env vars are permitted as a transitional step).

Also, moved the actual Ginkgo wiring logic out of GinkgoBootstrap; suites
should do it themselves as they may each have different configuration (i.e.
reporters).

Note that this currently creates and destroys the org on every example. Once
Before/AfterSuite are here we'll switch to that. Crank up them -nodes.

Signed-off-by: Aram Price <aprice@pivotallabs.com>