Adding Test Case for memory quota management
As quota management is a very important feature of CF so there should be a test case in CATS to validate this feature.

If a user pushes an app which needs a memory quota more than that of organization memory quota then it should fail while starting the app.

A similar type of test case for this is already there in (diego-acceptance-tests)[https://github.com/cloudfoundry-incubator/diego-acceptance-tests/blob/master/diego/staging_failure_test.go#L45]