Add JSON-formatted environment endpoint to Dora
The current `/env` endpoint returns a serialized Ruby hash. For this
response to be manipulated programmatically, it must be interpreted
back into a hash with a Ruby `eval`, which requires a Ruby interpreter
and is demonstrably insecure. A variant on this endpoint that returns
the environment as a JSON object is much safer and is language-agnostic.

There are no doubt some CATs that could improve the robustness of
their assertions by manipulating this JSON-formatted response.

No tests, but the Dora README is updated.