Do not use gzip -k option for compatibility
Older versions of gzip =< 1.5 do not support the -k option.

This is a problem, for instance,  when running the cf-acceptance-tests
using a golang docker image based on debian wheezy, which bundles gzip 1.5.

Here we refactor the test to uncompress and compress again the files,
simulation the behaviour of gzip -k. This change should be compatible
across multiple OS versions.