v3 helpers retrieves doppler endpoint from v2/info
The v3 helper has building the doppler URL in a hardcoded way. It does
not include the protocol, so it assumes HTTP, does not add the port and
assumes that the name is s/api/doppler/

In this commit we query the /v2/info endpoint to extract the doppler
endpoint. We replace the websocket protocol (ws) with HTTP (http). If
it is a SSL Websocket (wss) the string replacement will result in https.