Increase large payload test from 5kB to 200kB
I can no longer see the referenced story, but from 34d1c48 it looks like the
original intention of this test was highlight problems delivering large
responses from CF applications caused by MTU misconfiguration.

We recently experienced a problem like this which has been described in
cloudfoundry/guardian#77. This test passed because the response size of 5kB
was smaller than the MTU size of 9kB on some AWS instance types:

- http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/network_mtu.html#jumbo_frame_instances

Increasing this to a larger value will make the test more effective. I've
chosen a value of 200kB because it's larger than standard MTUs and
indicative of the response from an average website.