[#149872014] Improve security group test logic
The test used to check for the return code of curl.  This was
insufficient, as curl uses a return code of 7 for a variety of
connection failures.

We've improved the test to check for precisely what it needs: fail if
you get a "connection refused" after binding a security group to the
app, and fail if you *don't* get a "connection refused" after unbinding.