Test service instance sharing (#255)
* Add test of service instance sharing

The new test validates that one user can share a service instance into
a different space. A regular user who only has access to the target space
should be able to view the service when they call `cf services`. Later,
functionality will be added to allow the user in the target space to be
able to bind the shared service to an app.

Because this functionality is only available if the
service_instance_sharing feature flag is enabled, this commit also adds
a new integration config option for turning this test on or off.

[#150801597]

Signed-off-by: Sam Gunaratne <sgunaratne@pivotal.io>

* Modify service instance sharing CAT to test service binding

[#150801602]

* Add tests for unsharing service instance

[#150801613]

Signed-off-by: Sam Gunaratne <sgunaratne@pivotal.io>

* Test unsharing even if binding exists in target space

[#150801615]

Signed-off-by: Denise Yu <dyu@pivotal.io>

* Change Eventually to Wait

* Eventually defaults to 1s timeout. We want to use the
Config.DefaultTimeout

[#150801615]

Signed-off-by: Denise Yu <dyu@pivotal.io>

* Test unbinding of shared service instances

[#150801612]

Signed-off-by: Sam Gunaratne <sgunaratne@pivotal.io>

* fix changes introduced by new cli version

Signed-off-by: Sam Gunaratne <sgunaratne@pivotal.io>


[Finishes #153251152]