Fix/un-disable the credhub service key test
It appears that it was failing because of a change made in e28bfaed9e where the
broker was changed to create the credential and assign permissions
separately. Credhub appears to fail when you try to add permissions to a
credential that you already have permissions for.

Easiest fix was just to stop creating permissions when the bind is
being made for a service key, which we can tell by the app guid being
empty.