Windows metrics test matches linux test
The existing Windows metrics test tested that there were metrics emitted
to loggregator that had the origin "garden-windows". This change removes
this test as it is not a real "acceptance test" for CF and also becomes
incompatible once windows2012R2 cells disable UDP on the metron agent. We have
matched the `apps/loggregator` metrics tests in order to get parity
between the stacks.

Signed-off-by: Sunjay Bhatia <sbhatia@pivotal.io>