Preserve CF_PLUGIN_HOME for the cats run
We'd like to preserve plugin directory, so that they are shared between ginkgo nodes.
This is especially important now that we need log-cache CLI by default to be able to get
logs for apps. 'run-cats' task in cf-deployment-concourse-tasks is already doing it, but we're
adding it here so that we can run tests locally