Make windows/worker.go test more generic about log lines it expects
Now that we are using log-cache by default, logs are delivered faster sometimes. This means
that expectations in this test are two specific, as sometimes we get logs starting with line
'Running Worker 11'. This change asserts, more generically, that we first see a single digit,
and then a double digit.

[#160188077]