Fix failing ASG/C2C CAT on newer rootfs
Failing `[security_groups] App Instance Networking Using container-networking and running security-groups [It] correctly configures asgs and c2c policy independent of each other`

With a newer rootfs, the `ip` command now includes a `uid 2000` at the
end which breaks the `/myip` endpoint on the test app. This updated
command should be compatible with both the old and new `ip` output.