Restage app after applying manifest to get correct stack
This started failing after the default stack switched to
cflinuxfs3 in cf-deployment and we realized that
the CLI is returning the stack from the previously
staged droplet in the `cf app APP_NAME` command.

We need to restage the app to update the current droplet
with the manifest-assigned stack.

We removed the process command test from this particular
test since it is tested elsewhere in this file and
the dora test app cannot restage successfully with it.

See the following CLI Issue for more information:
https://github.com/cloudfoundry/cli/issues/1545

[finishes #163210854]

Signed-off-by: Tim Downey <tdowney@pivotal.io>