Improve error reporting when API redirects multiple times
'can copy package bits to another app and download the package' fails with a confusing unzip error, when api redirects multiple times, because it tries to unzip the body of a redirect. This change stops on the second redirect with an error message emulating what the cf CLI would do.