assets/nora: Add binaries required by the nora app
These binaries are required to build the Nora app which depends on
them to run tests when wats (Windows Acceptance Tests) are run.

The NoraPublished/ dir is a published (compiled) version of the app,
and the Nora/ dir is the unpublished source code which a user can
modify, compile and use for cats/other tests.

`bin/BreakoutBomb.exe` is an empty file maintained for legacy reasons

[#160865997]

Signed-off-by: Arjun Sreedharan <asreedharan@pivotal.io>