Use https syslog drain URL instead of syslog-tls.
Otherwise CATs will fail to run on certain configuration. In particular, the
`Syslog Drains` suite [1] will fail when `enable_unproxied_port_mappings` [2]
is set to `false`. I think the reasons are:

1. Based on the comments in CATs [3] and the conversations I had with the
networking team.  On a vanilla CF configured to use the silk CNI, the
syslog-agent cannot connect to the container's internal IP and, therefore, must
use the external IP.

2. When the above Diego flag is set to false, Diego will only map the envoy
ports to host ports.  This change requires running CATs with the
`require_proxied_app_traffic` flag set to true, changing the test's behavior to
use syslog-tls instead of syslog.

3. Diego container proxy certificate contains the internal IP in the SAN but
not the external IP.  This will cause the syslog-agent to fail to connect to
the syslog drain causing the test to fail.

1: https://github.com/cloudfoundry/cf-acceptance-tests/blob/82adc190668f61afa69aa1c6c66252493b2b5797/apps/syslog_drain.go#L91
2: https://github.com/cloudfoundry/diego-release/blob/2e6ff3c02f457c9daefa1de3df202a2986fdd8db/jobs/rep/spec#L287
3: https://github.com/cloudfoundry/cf-acceptance-tests/blob/82adc190668f61afa69aa1c6c66252493b2b5797/apps/syslog_drain.go#L92-L96

Authored-by: John Shahid <jshahid@pivotal.io>