service keys: allow error message on STDERR
When no service key is found, CLI v7 and earlier printed an error
message to STDOUT. v8 and above with print it to STDERR as this is the
standard for refactored commands. The test has been update to accept
either.

Additionally a lint error with the name of ASYNC_OPERATION_POLL_INTERVAL
was fixed by changing it to asyncOperationPollInterval, and an unchecked
error is now checked.