lifecycle tests: accept different error in test for contour
- When an app is unreachable, curling a route when we are using Istio as
our control plane results in a 404. However, when we are using Contour,
curling a route causes the connection to be closed and curl to hang.
This was causing our lifecycle tests to fail since it expects an error
to be returned.
- We are temporarily fixing the issue by checking if we are using
contour and if we are we check to see if curl failed with `exit code 35`
which is the error return by curl.

Co-authored-by: Leah Hanson <lhanson@pivotal.io>