Drop -d flag no longer supported after cf6
This script used the now-defunct '-d' flag for cf push, which was deprecated in version 7 of the cf CLI. It is no longer possible to specify an app's domain via commandline flags in a 'cf push', and so the script has been updated to retrieve the route of the test broker after it has been deployed. It then uses this to register the broker.