Fix security group failing encoding
`cf create-security-group ...` was failing in CATs in the security_groups
test suite inside of the `createSecurityGroup` function. The error
read:
```
Rules[0]: ports must be a valid single port, comma separated list of ports, or range or ports, formatted as a string
FAILED
```

Therefore it did not seem that the `allowedDestinations` struct was
being encoded correctly. When I altered the Destination struct to remove
the unnecessary "string" tag, my local testing started to work.