Test new /v3/service_instances/:guid/permissions
Also update the CATS service broker to retrieve the value of the HTTP_X_API_INFO_LOCATION header from calls from the Cloud Controller, store it in memory, and then serve it on a new endpoint. This is a slightly more realistic way to test the SSO lifecycle, as the old helper function used by these tests hardcoded the API info endpoint as `/info`. That worked because, like `/v2/info`, its response object also had the authorization_endpoint at the top level.