Fix CatsConfig and config
Due to [changes in cf-test-helpers](https://github.com/cloudfoundry/cf-test-helpers/pull/92),
we need to update our interface and interface implementation to include
some new methods.

As no one's asking for multiple origin support in CATs, I didn't
actually make it available via the integration_config. That's a trivial
change if we want to add it in the future.