Refactors credhub-service broker for clarity
Some differences in the way the new app works:
* Shares service UUID and plan UUID between catalog calls, rather than
  generating each time.
* Logs and returns internal server errors on certain failures.
* Only spins up credhub client on startup and reuses throughout.
* No manifest, as the old one was just setting a GOPACKAGENAME, which
  isn't necessary anymore.

In particular, that second bullet is important, as there were some
consistent flakes when running CATs due to a one-time failure to connect
with credhub that wasn't handled appropriately by the old app, and
resulted in a panic.

Also updates the tests to use this new app more appropriate, with the
`--no-start` flag and then starting it up after environment variables
are set. This should actually speed up runtime!