Combine worker assets
We currently maintain two Golang "worker" assets, one which is
ostensibly for Windows, and another which... isn't?

It seems like they both do relatively the same thing and that there's no
need to differentiate them by OS. So in the interest of minimizing our
assets footprint I deleted assets/worker-app and updated the relevant
tests to use assets/worker.