fix: Delete service brokers after tests
Fixes an inconsistency in the credhub, docker, and route_services test
suites (vs other suites) where apps were not get deleted after tests.
Specifically, some service brokers that were pushed as apps were not
getting deleted, only unregistered as service brokers.

I ran through some of the steps manually and saw that the
delete-service-broker command did not also delete the service broker
app.

Spot-checking the credhub suite, I couldn't find any reasons in git or
tracker stories for the delete command to not to be originally included.
Possibly, it was believed that the delete-service-broker command would
delete the app...