CI: Add GH actions workflow (#1371)
* CI: Add GH actions workflow

* CI: Remove schedule from CI workflow

Removes the unnecessary schedule from the GH Actions CI workflow. The
schedule is unnecessary because there's currently no external dependency
in any of the jobs in the workflow that could benefit from testing over
time, without code changes. For example, that usually takes the form of
a linting job, where the linting tool improves over time.