fix: Bump the minimum required version of Go to go1.23.0
`go mod tidy` keeps trying to change the minimum version of Go in the
go.mod. Looking it up, it seems like that's probably due to a dependency
requiring a higher minimum version of Go than what's currently set,
go1.22. We should make this change then.

Also, updates the toolchain to be go1.24.0. Left it in, but not sure we
care too much what this is set to.